Game works in console, mostly
Game is played manually entirely in console. Wrong winner name may be
declared. Need to decide if game play functions should go under
createPlayer or displayController (right now it's on displayController
and is commented out in createPlayer). Will try to make it really
modular as per the Modular Javascript #2 Youtube video given in the
lesson.